Code Analysis: [VS2019] Fixed reported issues

+++ GunBox/Sources/Engine/Sources/Application/CoreApplication.cpp
@@ -120,7 +120,7 @@ CoreApplication::Initialize()
     if (commandLineArgs->Renderer()) {
       auto renderer = commandLineArgs->Renderer().value();
 
-      LogInfo("Desired renderer: %d", renderer);
+      LogInfo("Desired renderer: %d", static_cast<int>(renderer));
 
       creationPreferences.RendererAPI() = renderer;
     }
